#0: Refactor Resnet files to match other model's folder/file structure and add tg, tgg tests/results
