+++ models/demos/t3000/resnet50/README.md
@@ -11,16 +11,10 @@ ResNet50 is a deep convolutional neural network architecture with 50 layers, des
 + The model picks up certain configs and weights from TorchVision pretrained model. We have used `torchvision.models.ResNet50_Weights.IMAGENET1K_V1` version from TorchVision as our reference.
 + Our ImageProcessor on the other hand is based on `microsoft/resnet-50` from huggingface.
 
-Please refer to the following directories for device specific versions and tests
+## Performance
 
-### Single Device
-[Grayskull](/models/demos/grayskull/resnet50/)
-
-[Wormhole_B0](/models/demos/wormhole/resnet50/)
-
-### Multi Device
-[Wormhole_B0 T3000](/models/demos/t3000/resnet50/)
-
-[Wormhole_B0 TG](/models/demos/tg/resnet50/)
-
-[Wormhole_B0 TGG](/models/demos/tgg/resnet50/)
+### T3000
+#### End-to-End Performance
++ For end-to-end performance, run `WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest models/demos/t3000/resnet50/tests/test_perf_e2e_resnet50.py::test_perf_trace_2cqs`.
++ This will generate a CSV with the timings and throughputs.
++ **Expected end-to-end perf**: For batch = 16 per device, or batch 128 in total, it is about `31,250 fps` currently. This may vary machine to machine.
