+++ models/demos/ttnn_resnet/tests/resnet50_performant.py
@@ -8,8 +8,7 @@ import ttnn
 from models.utility_functions import (
     is_wormhole_b0,
 )
-from models.utility_functions import run_for_wormhole_b0, run_for_grayskull
-from models.demos.ttnn_resnet.tests.ttnn_resnet_test_infra import create_test_infra
+from models.demos.ttnn_resnet.tests.resnet50_test_infra import create_test_infra
 
 try:
     from tracy import signpost
@@ -19,9 +18,7 @@ except ModuleNotFoundError:
     use_signpost = False
 
 
-def run_resnet50_inference(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
+def run_resnet50_inference(device, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator):
     if batch_size == 8:
         pytest.skip("Skipping batch size 8 due to memory config issue")
     if is_wormhole_b0() and batch_size == 20:
@@ -60,20 +57,16 @@ def run_resnet50_inference(
 
 def run_resnet50_trace_inference(
     device,
-    use_program_cache,
     batch_size,
     act_dtype,
     weight_dtype,
     math_fidelity,
-    enable_async,
     model_location_generator,
 ):
     if batch_size == 8:
         pytest.skip("Skipping batch size 8 due to memory config issue")
     if is_wormhole_b0() and batch_size == 20:
         pytest.skip("Skipping batch size 20 for Wormhole B0 due to fitting issue")
-
-    device.enable_async(enable_async)
 
     test_infra = create_test_infra(
         device,
@@ -126,10 +119,10 @@ def run_resnet50_trace_inference(
         signpost(header="stop")
     test_infra.validate()
 
+    ttnn.release_trace(device, tid)
 
-def run_resnet50_2cqs_inference(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
+
+def run_resnet50_2cqs_inference(device, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator):
     if batch_size == 8:
         pytest.skip("Skipping batch size 8 due to memory config issue")
     if is_wormhole_b0() and batch_size == 20:
@@ -192,20 +185,16 @@ def run_resnet50_2cqs_inference(
 
 def run_resnet50_trace_2cqs_inference(
     device,
-    use_program_cache,
     batch_size,
     act_dtype,
     weight_dtype,
     math_fidelity,
-    enable_async,
     model_location_generator,
 ):
     if batch_size == 8:
         pytest.skip("Skipping batch size 8 due to memory config issue")
     if is_wormhole_b0() and batch_size == 20:
         pytest.skip("Skipping batch size 20 for Wormhole B0 due to fitting issue")
-
-    device.enable_async(enable_async)
 
     test_infra = create_test_infra(
         device,
@@ -291,179 +280,3 @@ def run_resnet50_trace_2cqs_inference(
         test_infra.validate(output)
 
     ttnn.release_trace(device, tid)
-
-
-@run_for_grayskull()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((20, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-def test_run_resnet50_inference_gs(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
-    run_resnet50_inference(
-        device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-    )
-
-
-@run_for_grayskull()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768, "trace_region_size": 1332224}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((20, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-@pytest.mark.parametrize("enable_async", [True, False])
-def test_run_resnet50_trace_inference_gs(
-    device,
-    use_program_cache,
-    batch_size,
-    act_dtype,
-    weight_dtype,
-    math_fidelity,
-    enable_async,
-    model_location_generator,
-):
-    run_resnet50_trace_inference(
-        device,
-        use_program_cache,
-        batch_size,
-        act_dtype,
-        weight_dtype,
-        math_fidelity,
-        enable_async,
-        model_location_generator,
-    )
-
-
-@run_for_grayskull()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768, "num_hw_cqs": 2}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((20, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-def test_run_resnet50_2cqs_inference_gs(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
-    run_resnet50_2cqs_inference(
-        device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-    )
-
-
-@run_for_grayskull()
-@pytest.mark.parametrize(
-    "device_params", [{"l1_small_size": 32768, "trace_region_size": 1332224, "num_hw_cqs": 2}], indirect=True
-)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((20, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-@pytest.mark.parametrize("enable_async", [True, False])
-def test_run_resnet50_trace_2cqs_inference_gs(
-    device,
-    use_program_cache,
-    batch_size,
-    act_dtype,
-    weight_dtype,
-    math_fidelity,
-    enable_async,
-    model_location_generator,
-):
-    run_resnet50_trace_2cqs_inference(
-        device,
-        use_program_cache,
-        batch_size,
-        act_dtype,
-        weight_dtype,
-        math_fidelity,
-        enable_async,
-        model_location_generator,
-    )
-
-
-@run_for_wormhole_b0()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 24576}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((16, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-def test_run_resnet50_inference_wh(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
-    run_resnet50_inference(
-        device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-    )
-
-
-@run_for_wormhole_b0()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 24576, "trace_region_size": 800768}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((16, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-@pytest.mark.parametrize("enable_async", [True, False])
-def test_run_resnet50_trace_inference_wh(
-    device,
-    use_program_cache,
-    batch_size,
-    act_dtype,
-    weight_dtype,
-    math_fidelity,
-    enable_async,
-    model_location_generator,
-):
-    run_resnet50_trace_inference(
-        device,
-        use_program_cache,
-        batch_size,
-        act_dtype,
-        weight_dtype,
-        math_fidelity,
-        enable_async,
-        model_location_generator,
-    )
-
-
-@run_for_wormhole_b0()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 24576, "num_hw_cqs": 2}], indirect=True)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((16, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-def test_run_resnet50_2cqs_inference_wh(
-    device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-):
-    run_resnet50_2cqs_inference(
-        device, use_program_cache, batch_size, act_dtype, weight_dtype, math_fidelity, model_location_generator
-    )
-
-
-@run_for_wormhole_b0()
-@pytest.mark.parametrize(
-    "device_params", [{"l1_small_size": 24576, "trace_region_size": 800768, "num_hw_cqs": 2}], indirect=True
-)
-@pytest.mark.parametrize(
-    "batch_size, act_dtype, weight_dtype, math_fidelity",
-    ((16, ttnn.bfloat8_b, ttnn.bfloat8_b, ttnn.MathFidelity.LoFi),),
-)
-@pytest.mark.parametrize("enable_async", [True, False])
-def test_run_resnet50_trace_2cqs_inference_wh(
-    device,
-    use_program_cache,
-    batch_size,
-    act_dtype,
-    weight_dtype,
-    math_fidelity,
-    enable_async,
-    model_location_generator,
-):
-    run_resnet50_trace_2cqs_inference(
-        device,
-        use_program_cache,
-        batch_size,
-        act_dtype,
-        weight_dtype,
-        math_fidelity,
-        enable_async,
-        model_location_generator,
-    )
