+++ pytest.ini
@@ -14,4 +14,6 @@ markers =
     models_performance_virtual_machine: mark model silicon tests for performance on virtual_machine
     models_device_performance_bare_metal: mark model silicon tests for device performance on bare metal
     model_perf_t3000: mark model silicon tests for performance on t3000 bare metal
+    model_perf_tg: mark model silicon tests for performance on tg bare metal
+    model_perf_tgg: mark model silicon tests for performance on tgg bare metal
     requires_fast_runtime_mode_off: mark tests which require fast runtime mode to be off: validation, logging, tracing
