+++ tests/nightly/wh_b0_only_eth/models/demos/wormhole/resnet50/tests/test_resnet50_performant.py
@@ -1,0 +1,1 @@
+../../../../../../../../models/demos/wormhole/resnet50/tests/test_resnet50_performant.py
