+++ tests/scripts/single_card/nightly/run_gs_only.sh
@@ -10,7 +10,8 @@ fail=0
 
 echo "Running model nightly tests for GS only"
 
-env pytest -n auto models/demos/ttnn_resnet/tests/test_ttnn_resnet50_performant.py ; fail+=$?
+env pytest -n auto models/demos/grayskull/resnet50/tests/test_resnet50_performant.py ; fail+=$?
+env pytest -n auto models/demos/grayskull/resnet50/demo/demo.py ; fail+=$?
 
 if [[ $fail -ne 0 ]]; then
   exit 1
