+++ tests/scripts/t3000/run_t3000_frequent_tests.sh
@@ -131,7 +131,7 @@ run_t3000_resnet_tests() {
 
   echo "LOG_METAL: Running run_t3000_resnet_tests"
 
-  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto models/demos/ttnn_resnet/tests/multi_device/test_ttnn_resnet50_performant.py ; fail+=$?
+  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto models/demos/t3000/resnet50/tests/test_resnet50_performant.py ; fail+=$?
 
   # Record the end time
   end_time=$(date +%s)
