+++ tests/scripts/tg/run_tg_model_perf_tests.sh
@@ -11,6 +11,10 @@ run_tg_llm_tests() {
 }
 
 run_tg_cnn_tests() {
+
+  echo "LOG_METAL: Running run_tg_resnet50_tests"
+  env pytest -n auto models/demos/tg/resnet50/tests/test_perf_e2e_resnet50.py -m "model_perf_tg" ; fail+=$?
+
   # Merge all the generated reports
   env python models/perf/merge_perf_results.py; fail+=$?
