+++ tests/ttnn/integration_tests/resnet/test_ttnn_functional_resnet50_xxlarge_new.py
@@ -23,7 +23,7 @@ from models.utility_functions import (
     is_wormhole_b0,
 )
 
-from models.demos.ttnn_resnet.tests.ttnn_resnet_test_infra import load_resnet50_model
+from models.demos.ttnn_resnet.tests.resnet50_test_infra import load_resnet50_model
 from models.demos.ttnn_resnet.tt.ttnn_functional_resnet50_xxlarge_new_conv_api import resnet50
